Successfuly delete now sends back deleted item

+++ index.js
@@ -29,7 +29,7 @@ server.get('/api/users/:id', (req, res) => {
 server.post('/api/users', (req, res) => {
     db.insert(req.body)
     .then(user => {
-        res.status(200).json(user);
+        res.status(200).json(req.body);
     })
     .catch(err => {
 
@@ -38,13 +38,21 @@ server.post('/api/users', (req, res) => {
 
 server.delete('/api/users/:id', (req, res) => {
     const { id } = req.params;
-    db.remove(id)
-    .then(deleted => {
-        res.status(204).end();
+    db.findById(id)
+    .then(user => {
+        if (user) {
+            db.remove(id)
+            .then(deleted => {
+                res.status(201).json(user)
+            })
+            .catch(err => {
+                res.status(501).json({message: 'Server error'})
+            })
+        } else {
+            res.status(500).json({message: 'Something went wrong'})
+        }
     })
-    .catch(err => {
-
-    })
+    .catch(err => res.status(404).json({message: 'User not found'}))
 })
 
 server.put('/api/users/:id', (req, res) => {
@@ -53,7 +61,7 @@ server.put('/api/users/:id', (req, res) => {
     db.update(id, userInfo)
     .then(updated => {
         if (updated) {
-            res.status(200).json(updated)
+            res.status(200).json(userInfo)
         } else {
             res.status(404).json({ message: "User not found"})
         }
